Rename config encryption filter from config-secure.
