+++ bom/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (c) 2017, 2018 Oracle and/or its affiliates. All rights reserved.
+    Copyright (c) 2017, 2019 Oracle and/or its affiliates. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -101,7 +101,7 @@
             </dependency>
             <dependency>
                 <groupId>io.helidon.config</groupId>
-                <artifactId>helidon-config-secure</artifactId>
+                <artifactId>helidon-config-encryption</artifactId>
                 <version>${project.version}</version>
             </dependency>
             <!-- security -->
