+++ bundles/security/src/main/java9/module-info.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
+ * Copyright (c) 2018,2019 Oracle and/or its affiliates. All rights reserved.
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -20,7 +20,7 @@
 module io.helidon.security.bundle {
     requires transitive io.helidon.security;
     requires transitive io.helidon.security.annotations;
-    requires transitive io.helidon.config.secure;
+    requires transitive io.helidon.config.encryption;
     requires transitive io.helidon.security.providers.header;
     requires transitive io.helidon.security.providers.httpauth;
     requires transitive io.helidon.security.providers.httpsign;
