+++ config/encryption/src/test/resources/application.yaml
