+++ config/pom.xml
@@ -1,7 +1,7 @@
 <?xml version="1.0" encoding="UTF-8"?>
 <!--
 
-    Copyright (c) 2017, 2018 Oracle and/or its affiliates. All rights reserved.
+    Copyright (c) 2017, 2019 Oracle and/or its affiliates. All rights reserved.
 
     Licensed under the Apache License, Version 2.0 (the "License");
     you may not use this file except in compliance with the License.
@@ -42,7 +42,7 @@
         <module>etcd</module>
         <module>git</module>
         <module>hocon</module>
-        <module>secure</module>
+        <module>encryption</module>
         <module>testing</module>
         <module>test-infrastructure</module>
         <module>tests</module>
