--- config/secure/src/main/java/io/helidon/config/secure/SecureConfigFilterService.java
@@ -1,41 +1,0 @@
-/*
- * Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.helidon.config.secure;
-
-import io.helidon.config.Config;
-import io.helidon.config.spi.ConfigFilter;
-
-/**
- * A Java service for {@link SecureConfigFilter}.
- */
-public class SecureConfigFilterService implements ConfigFilter {
-    private ConfigFilter filter;
-
-    @Override
-    public String apply(Config.Key key, String stringValue) {
-        if (null == filter) {
-            return stringValue;
-        }
-
-        return filter.apply(key, stringValue);
-    }
-
-    @Override
-    public void init(Config config) {
-        this.filter = SecureConfigFilter.fromConfig().apply(config);
-    }
-}
