--- config/secure/src/test/java/io/helidon/config/secure/TestFilterService.java
@@ -1,38 +1,0 @@
-/*
- * Copyright (c) 2018 Oracle and/or its affiliates. All rights reserved.
- *
- * Licensed under the Apache License, Version 2.0 (the "License");
- * you may not use this file except in compliance with the License.
- * You may obtain a copy of the License at
- *
- *     http://www.apache.org/licenses/LICENSE-2.0
- *
- * Unless required by applicable law or agreed to in writing, software
- * distributed under the License is distributed on an "AS IS" BASIS,
- * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- * See the License for the specific language governing permissions and
- * limitations under the License.
- */
-
-package io.helidon.config.secure;
-
-import io.helidon.config.Config;
-
-import org.junit.jupiter.api.Test;
-
-import static org.hamcrest.CoreMatchers.is;
-import static org.hamcrest.MatcherAssert.assertThat;
-
-/**
- * Test for automatic filter loading.
- */
-class TestFilterService {
-    @Test
-    void testFiltering() {
-        Config config = Config.create();
-
-        String value = config.get("pwd3").asString().get();
-
-        assertThat(value, is(AbstractSecureConfigTest.TEST_STRING));
-    }
-}
